+++ frontend/src/components/Category.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import PostBody  from './PostBody';
+import NavDropdownMenu from './NavDropdownMenu';
+
+export default class Category extends React.Component {
+    render() {
+      const { match } = this.props
+      const category = match.params.categoryPath
+  
+      return (
+        <div>
+            <NavDropdownMenu />
+            <PostBody title={category} category={category} />
+        </div>
+      )
+    }
+  }
