+++ frontend/src/components/index.js
@@ -3,6 +3,8 @@ export * from './Home'
 export * from './VoteUpAndDown'
 export * from './PostCardList'
 export * from './PostCard'
+export * from './PostBody'
 export * from './FilterSelect'
 export * from './AddOrEditPost'
 export * from './App'
+export * from './Category'
